pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "travel-planner"
version = "0.1.0"
description = "Agentic travel planner - PR-1 contracts and foundations"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0",
    "python-dateutil>=2.8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "types-python-dateutil",
    "hypothesis>=6.0.0",
    "pre-commit>=3.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*", "eval*", "scripts*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
